♻️ refactor(config): 重构配置管理器，采用服务拆分模式
- 【重构】将配置管理器拆分为多个独立服务以提高代码组织和维护性：
  - ConfigurationService：处理核心配置读写操作
  - ConfigurationChangeHandler：处理配置变更事件
  - ProviderConfigValidator：验证AI提供商配置
  - ConfigurationMonitor：监控配置变更和系统响应
- 【优化】简化配置管理器主类，将复杂逻辑迁移至专门的服务类
- 【改进】增加每个服务模块的详细文档说明
- 【重构】优化配置变更处理机制，提高代码复用性
- 【修改】采用依赖注入模式重组服务间关系，降低耦合度
